refactor: added a util for similar code
