+++ src/forms/login-popup/index.jsx
@@ -16,7 +16,6 @@ import { loginUser, setLoginSSOIntent } from './data/reducers';
 import messages from './messages';
 import { InlineLink, SocialAuthProviders } from '../../common-ui';
 import {
-  AUTH_MODE,
   AUTH_MODE_LOGIN,
   COMPLETE_STATE,
   ENTERPRISE_LOGIN_URL,
@@ -27,7 +26,7 @@ import {
   TPA_AUTHENTICATION_FAILURE,
 } from '../../data/constants';
 import { useDispatch, useSelector } from '../../data/storeHooks';
-import getAllPossibleQueryParams, { moveScrollToTop } from '../../data/utils';
+import getAllPossibleQueryParams, { handleURLUpdationOnLoad, moveScrollToTop } from '../../data/utils';
 import { setCurrentOpenedForm } from '../../onboarding-component/data/reducers';
 import {
   trackForgotPasswordLinkClick, trackLoginPageViewed, trackLoginSuccess, trackRegisterFormToggled,
@@ -103,14 +102,7 @@ const LoginForm = () => {
     if (moveScrollToTop) {
       moveScrollToTop(loginFormHeadingRef, 'end');
     }
-    const queryParam = getAllPossibleQueryParams();
-    if (!Object.prototype.hasOwnProperty.call(queryParam, AUTH_MODE)
-      || queryParam?.[AUTH_MODE] !== AUTH_MODE_LOGIN) {
-      const url = new URL(window.location.href);
-      url.searchParams.delete(AUTH_MODE);
-      url.searchParams.set(AUTH_MODE, AUTH_MODE_LOGIN);
-      window.history.replaceState(null, null, url);
-    }
+    handleURLUpdationOnLoad(AUTH_MODE_LOGIN);
   }, []);
 
   useEffect(() => {
