+++ src/forms/registration-popup/index.jsx
@@ -22,7 +22,6 @@ import isFormValid from './data/utils';
 import messages from './messages';
 import { InlineLink, SocialAuthProviders } from '../../common-ui';
 import {
-  AUTH_MODE,
   AUTH_MODE_REGSITER,
   COMPLETE_STATE,
   ENTERPRISE_LOGIN_URL,
@@ -33,7 +32,9 @@ import {
   TPA_AUTHENTICATION_FAILURE,
 } from '../../data/constants';
 import { useDispatch, useSelector } from '../../data/storeHooks';
-import getAllPossibleQueryParams, { getCountryCookieValue, moveScrollToTop, setCookie } from '../../data/utils';
+import getAllPossibleQueryParams, {
+  getCountryCookieValue, handleURLUpdationOnLoad, moveScrollToTop, setCookie,
+} from '../../data/utils';
 import { setCurrentOpenedForm } from '../../onboarding-component/data/reducers';
 import './index.scss';
 import {
@@ -144,14 +145,7 @@ const RegistrationForm = () => {
 
   useEffect(() => {
     moveScrollToTop(registerFormHeadingRef, 'end');
-    const queryParam = getAllPossibleQueryParams();
-    if (!Object.prototype.hasOwnProperty.call(queryParam, AUTH_MODE)
-      || queryParam?.[AUTH_MODE] !== AUTH_MODE_REGSITER) {
-      const url = new URL(window.location.href);
-      url.searchParams.delete(AUTH_MODE);
-      url.searchParams.set(AUTH_MODE, AUTH_MODE_REGSITER);
-      window.history.replaceState(null, null, url);
-    }
+    handleURLUpdationOnLoad(AUTH_MODE_REGSITER);
   }, []);
 
   useEffect(() => {
